+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Field Types - Support `TIMESTAMP(n)`, `SMALLINT(n)` and `BIGINT(n)` column types (MySQL)
 
+### Fixed
+- Command Generate - Generate MongoDB belongsTo.
+
 ## RELEASE 2.6.3 - 2019-11-04
 ### Fixed
 - Command Generate - Fix generation of project with a mongo database.